+++ p2panda-stream/src/dependencies/mod.rs
@@ -1,0 +1,5 @@
+mod checker;
+mod store;
+
+pub use checker::{DependencyChecker, DependencyCheckerError};
+pub use store::{DependencyStore, MemoryStore};
