--- p2panda-stream/src/ordering/mod.rs
@@ -1,2 +1,0 @@
-pub mod dependency_checker;
-pub mod operation_dependency_checker;
