+++ p2panda-stream/src/stream/mod.rs
@@ -2,6 +2,7 @@
 
 mod decode;
 mod ingest;
+mod dependencies;
 
 pub use decode::{Decode, DecodeExt};
 pub use ingest::{Ingest, IngestExt};
